feat: x range control
